+++ haystack_experimental/components/tools/openapi/__init__.py
@@ -2,7 +2,7 @@
 #
 # SPDX-License-Identifier: Apache-2.0
 
-from haystack_experimental.components.tools.openapi._openapi import LLMProvider
 from haystack_experimental.components.tools.openapi.openapi_tool import OpenAPITool
+from haystack_experimental.components.tools.openapi.types import LLMProvider
 
 __all__ = ["LLMProvider", "OpenAPITool"]
