+++ haystack_experimental/components/tools/openapi/openapi_tool.py
@@ -10,12 +10,12 @@ from haystack import component, logging
 from haystack.components.generators.chat import OpenAIChatGenerator
 from haystack.dataclasses import ChatMessage, ChatRole
 from haystack.lazy_imports import LazyImport
-from haystack_experimental.components.tools.openapi import LLMProvider
 
 from haystack_experimental.components.tools.openapi._openapi import (
     ClientConfiguration,
     OpenAPIServiceClient,
 )
+from haystack_experimental.components.tools.openapi.types import LLMProvider
 
 with LazyImport("Run 'pip install anthropic-haystack'") as anthropic_import:
     # pylint: disable=import-error
