+++ haystack_experimental/components/tools/openapi/types.py
@@ -1,0 +1,10 @@
+from enum import Enum
+
+
+class LLMProvider(Enum):
+    """
+    Enum for the supported LLM providers.
+    """
+    OPENAI = "openai"
+    ANTHROPIC = "anthropic"
+    COHERE = "cohere"
